refactor(datatable): move mapping helpers to mappings/ (refs #298)
