+++ src/datatable/dataHelpers.ts
@@ -9,7 +9,7 @@ import {
   TimeRange
 } from '@grafana/data';
 import { FormatColumnValue } from './cells/cellRenderer';
-import { ApplyGrafanaOverrides } from './overrides';
+import { ApplyGrafanaOverrides } from './mappings/overrides';
 import { CellMetaSettings, ConfigColumnDefs } from 'datatables.net';
 import { ColumnAlignment, ColumnAlignmentOptions, ColumnStyleColoring } from 'types';
 import { DTColumnType, FormattedColumnValue } from './types';
@@ -17,7 +17,7 @@ import { ColumnStyleItemType, ColumnStyles } from 'components/options/columnstyl
 import { ApplyColumnStyles } from './columns/columnStyles';
 import { DTData } from 'components/DataTablePanel';
 import { processRowColumnStyle, processRowStyle, ProcessStringValueStyle } from './cells/createdCellHelpers';
-import { ApplyMappings, GetMappings } from './mappingProcessor';
+import { ApplyMappings, GetMappings } from './mappings/mappingProcessor';
 
 function normalizeFieldName(field: string) {
   return field
